New script to produce "noncrop_precrop_mask"
This script is uses a new custom function (cc_extract_noncrop_precrop()) to produce a noncrop_precrop_mask.

This mask is used to remove all periods of noncropland that occurred before cultivation for each pixel. The mask is then used to mask the "max extent" max_abn_lcc_iucn obserd and potential habitat maps, which include habitat for each abandoned pixel for 1987-2017.

The masked versions are saved as .tif files and converted to data.tables (saved as .parquet files).